refactor: ReviewService 구조 원복
